# swagger.json
{"info":{"title":"Waffle server","version":"0.0.1"},"host":"waffle-server-dev.gapminderdev.org:443","schemes":["http","https"],"basePath":"/","swagger":"2.0","paths":{"/api/vizabi/translation/{lang}.json":{"get":{"description":"Translation file","produces":["application/json"],"parameters":[{"name":"lang","in":"path","required":true,"description":"specify the language","type":"string"}],"tags":["Translation"],"responses":{"200":{"description":"Translation file"},"304":{"description":"cache"},"default":{"description":"Unexpected error","schema":{"$ref":"#/definitions/Error"}}}}},"/api/vizabi/mc_precomputed_shapes.json":{"get":{"description":"mc_precomputed_shapes for mountain chart","produces":["application/json"],"tags":["McPecomputedShapes"],"responses":{"200":{"description":"McPecomputedShapes"},"304":{"description":"cache"},"default":{"description":"Unexpected error","schema":{"$ref":"#/definitions/Error"}}}}},"/api/vizabi/world-50m.json":{"get":{"description":"world-50m.json for bubbleMap chart","produces":["application/json"],"tags":["World-50m"],"responses":{"200":{"description":"world-50m.json"},"304":{"description":"cache"},"default":{"description":"Unexpected error","schema":{"$ref":"#/definitions/Error"}}}}},"/api/vizabi/metadata.json":{"get":{"description":"metadata.json for all charts","produces":["application/json"],"tags":["Metadata"],"responses":{"200":{"description":"metadata.json"},"304":{"description":"cache"},"default":{"description":"Unexpected error","schema":{"$ref":"#/definitions/Error"}}}}},"/api/vizabi/geo_properties.csv":{"get":{"description":"geo_properties.csv","produces":["text/csv"],"tags":["geo_properties"],"responses":{"200":{"description":"geo_properties.csv"},"304":{"description":"cache"},"default":{"description":"Unexpected error","schema":{"$ref":"#/definitions/Error"}}}}},"/api/graphs/stats/vizabi-tools":{"get":{"description":"For getting geo-props, geo-time, geo-time-measure","produces":["application/json","text/csv"],"parameters":[{"name":"select","in":"query","description":"array of columns.","type":"string"},{"name":"where","in":"query","description":"list of filters","type":"string"},{"name":"gapfilling","in":"query","description":"list of methods for post processing of measures data (isn't supported yet)","type":"string"}],"tags":["Graph"],"responses":{"200":{"description":"An array of products","schema":{"type":"array","items":{"$ref":"#/definitions/Graph"}}},"304":{"description":"cache","schema":{"type":"array","items":{"$ref":"#/definitions/Graph"}}},"default":{"description":"Unexpected error","schema":{"$ref":"#/definitions/Error"}}}}},"/api/geo":{"get":{"description":"Geo-poroperties","produces":["application/json","text/csv"],"parameters":[{"name":"select","in":"query","description":"array of columns. (by default following columns are used ['geo', 'geo.name', 'geo.cat', 'geo.region'] Order of columns is important)","type":"string"},{"name":"where","in":"query","description":"list of filters for dimensions","type":"string"},{"name":"gapfilling","in":"query","description":"list of methods for post processing of measures data (isn't supported yet)","type":"string"}],"tags":["Geo"],"responses":{"200":{"description":"An array of products","schema":{"type":"array","items":{"$ref":"#/definitions/Geo"}}},"304":{"description":"cache","schema":{"type":"array","items":{"$ref":"#/definitions/Geo"}}},"default":{"description":"Unexpected error","schema":{"$ref":"#/definitions/Error"}}}}}},"definitions":{"Graph":{"type":"object","properties":{"geo-props":{"type":"string","description":"Result will consist only from unique geo-props, filtered by geo values. Example query `?select=geo.latitude,geo.name,geo&geo.cat=country,region`"},"geo-time":{"type":"string","description":"Result will consist only from unique geo-time, filtered by time values. Example query `?select=geo,time`"},"geo-time-measure":{"type":"string","description":"Example query `?select=geo,time,population,gini`"},"geo-time-measure (filter by geo)":{"type":"string","description":"Example query `?select=geo,time,population,gini&geo=chn`"},"geo-time-measure (filter time)":{"type":"string","description":"Example query `?select=geo,time,population,gini&time=1800`, `?select=geo,time,population,gini&time=2000:2010`"}}},"Geo":{"type":"object","properties":{"geo":{"type":"string","description":"Result will consist only from unique geo-props, filtered by geo values."},"geo.name":{"type":"string","description":"Result will consist only from name country or region"},"geo.cat":{"type":"string","description":"May contain some or all of the given values - 'global,region,country,un_state,world_4region,geo'"},"geo.region":{"type":"string","description":"May contain some or all of the given value - 'asia,americas,europe,africa'"},"geo.latitude":{"type":"string","description":"Result will consist only from  coordinate (latitude) of the country"},"geo.longitude":{"type":"string","description":"Result will consist only from coordinate (longitude) of the country"}}},"Error":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"}}}}}